@@ -49,12 +51,12 @@ test {
 }
 
 jar {
-    baseName = 'scsb-etl'
-    version = '0.0.1-SNAPSHOT'
+    archiveBaseName = 'phase4-scsb-etl'
+    archiveVersion = '0.0.1-SNAPSHOT'
 }
 
-sourceCompatibility = 1.8
-targetCompatibility = 1.8
+sourceCompatibility = 1.11
+targetCompatibility = 1.11
 
 repositories {
     mavenCentral()
