@@ -1,7 +1,8 @@
 package org.recap.model.jaxb.marc;
 
-import org.apache.log4j.Logger;
+import org.slf4j.Logger;
 import org.recap.model.jaxb.JAXBContextHandler;
+import org.slf4j.LoggerFactory;
 
 import javax.xml.bind.Marshaller;
 import javax.xml.bind.Unmarshaller;
@@ -41,7 +42,7 @@ import java.util.List;
 @XmlRootElement(name = "collection")
 public class CollectionType implements Serializable {
 
-    private static final Logger LOG = Logger.getLogger(CollectionType.class);
+    private static final Logger LOG = LoggerFactory.getLogger(CollectionType.class);
 
     /**
      * The Record.
